Extract macro processing to parse_macro.go
